fix: tighten standard PostgreSQL task verifiers (#262)
