+++ tasks/postgres/standard/security/rls_business_access/description.md
@@ -18,13 +18,13 @@ Build RLS policies for a social platform where users create posts and comments i
 - **DELETE**: Only channel owners can delete channels
 
 ### 3. Posts Table Access Rules:
-- **SELECT**: Users can read all posts in channels they have access to
+- **SELECT**: Users can read all posts in channels they have access to. (Authors do NOT get a separate read privilege — visibility is determined solely by channel accessibility.)
 - **INSERT**: Authenticated users can create posts in any channel
 - **UPDATE**: Post authors OR channel moderators OR channel owners can edit posts
 - **DELETE**: Post authors OR channel moderators OR channel owners can delete posts
 
 ### 4. Comments Table Access Rules:
-- **SELECT**: Users can read comments on posts they can access
+- **SELECT**: Users can read comments on posts they can access. (Comment authors do NOT get a separate read privilege — visibility follows the post's channel accessibility only.)
 - **INSERT**: Authenticated users can comment on posts they can see
 - **UPDATE**: Comment authors OR post authors OR channel moderators OR channel owners can edit comments
 - **DELETE**: Comment authors OR post authors OR channel moderators OR channel owners can delete comments
@@ -36,7 +36,7 @@ Build RLS policies for a social platform where users create posts and comments i
 
 ## Session Context:
 
-Use `current_setting('app.current_user_id')` to get the current user ID from session context.
+The session sets `app.current_user_id` to the user's UUID, or `''` for anonymous users. Use the pre-created helper `app_current_user_id()` in your policies — it returns the UUID or `NULL` for anonymous (a raw `::UUID` cast on the empty string would error).
 
 ## Schema Requirements:
 
@@ -48,11 +48,12 @@ Use `current_setting('app.current_user_id')` to get the current user ID from ses
 
 1. **Enable RLS** on all five tables
 2. **Create policies** for SELECT, INSERT, UPDATE, DELETE operations on each table
-3. **Helper functions** to check permissions efficiently:
+3. **Helper functions are pre-created** — use them in your policies:
+   - `app_current_user_id()` — returns the current user UUID (or `NULL` for anonymous)
    - `is_channel_owner(channel_id, user_id)`
    - `is_channel_moderator(channel_id, user_id)`
    - `can_moderate_channel(channel_id, user_id)`
-4. **Proper indexing** to ensure RLS policies perform well
+4. **Performance indexes are pre-created** — focus on writing correct, efficient policies.
 
 ## Test Scenarios:
